+++ WPF.Net.Examples/App.xaml.cs
@@ -1,12 +1,8 @@
 ﻿using System.Windows;
-// ReSharper disable UnusedMember.Global
 // ReSharper disable RedundantExtendsListEntry
 
 namespace WPF.Net.Examples
 {
-    /// <summary>
-    /// Interaction logic for App.xaml
-    /// </summary>
     public partial class App : Application
     {
         //