+++ WPF.Net.Examples/Properties/AssemblyInfo.cs
@@ -10,7 +10,7 @@ using System.Windows;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Damian Viktorovich Morozov")]
 [assembly: AssemblyProduct("WPF.Net.Examples")]
-[assembly: AssemblyCopyright("Copyright © 2020")]
+[assembly: AssemblyCopyright("Copyright (c) 2020")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
@@ -49,5 +49,5 @@ using System.Windows;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("0.2.41.0")]
-[assembly: AssemblyFileVersion("0.2.41.0")]
+[assembly: AssemblyVersion("0.2.44.0")]
+[assembly: AssemblyFileVersion("0.2.44.0")]