+++ WPF.Net.Examples/Views/PagePing.xaml.cs
@@ -55,7 +55,6 @@ namespace WPF.Net.Examples.Views
                     }
                 });
             }
-
         }
 
         private void PingWork()